+++ src/pages/namespace/Instances/Detail/Main/Diagram/index.tsx
