+++ src/pages/namespace/Instances/Detail/Main/InputOutput/OutputInfo.tsx
