+++ src/pages/namespace/Instances/Detail/Main/Logs/Entry.tsx
