+++ src/pages/namespace/Instances/Detail/Main/Logs/Filters/TextInput.tsx
