+++ src/pages/namespace/Instances/Detail/Main/Logs/utils.ts
